services/move_processed_files_to_cold_bucket_service/requirements.txt
google-cloud-bigquery==3.5.0
google-cloud-storage==2.10.0
cloudevents==1.9.0
fastapi==0.95.2
uvicorn==0.22.0
dacite==1.6.0
toolz==0.12.0